Support building on python-3.12
Loading ACT data is a requirement for the Simons Observatory, at
least for the next few years.  This means that moby2 is a required
piece of the SO software stack.  That in turn means that moby2 needs
to be installable in modern python environments.  This PR makes
several changes to support building with python-3.12.